+++ src/Services/Translation.php
@@ -12,6 +12,11 @@ class Translation
     protected static $config;
     
     /**
+     * @var string
+     */
+    protected static $prefix = '';
+    
+    /**
      * @param string $key
      * @param mixed $detault
      * @throws \Exception
@@ -21,19 +26,23 @@ class Translation
         if (is_null(self::$config)) {
             throw new \Exception('No data is loaded. Please use the load method');
         }
-        
+    
+        $key = !empty(self::$prefix) ? sprintf('%s.%s', self::$prefix, $key) : $key;
+    
         return self::$config->get($key, $default);
     }
     
     /**
      * @param string $json_file
+     * @param string $prefix
      */
-    public static function load($json_file)
+    public static function load($json_file, $prefix = '')
     {
         if (!file_exists($json_file)) {
             throw new \InvalidArgumentException(sprintf('File (%s) does not exists', $json_file));
         }
         
         self::$config = new Config($json_file);
+        self::$prefix = $prefix;
     }
 }