+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.0.2
 -----
 - Removed autoloader and debug stuff
+- Added key prefix
 
 0.0.1
 -----